Finally solved the color matching with Y-Axis
Finally figured out the best way to color match the Y-Axis. 
I created an array when the Y-Axis is drawn, and then use it later
to plot the data points. But I think I was still trying to figure out 
getting the data to plot in the graph area and not outside of it.