Move Colors to New File and add modopts.ini
This update is when I started changing a bit more of how things
worked and added a bit of customization that can be done in
modopts.ini. Specifically in modopts.ini you can now choose
what currency you want to have as your first listed exchange
rate, and also this will be the currency on the chart for historic.