Tried animation with event-timer.js
So the BitBomb animation worked, but on several terminals it
was causing some drawing errors. Specifically on Syncterm and
NetRunner I was seeing partial ANSI Escape Sequences on the 
terminal screen. I figured this was due to the speed of the characters
being drawn. Funny enough, this problem did not present itself
when using PuTTY set up with CP437 for character translation. 
In an attempt to slow down the animation, I tried using the
event-timer.js from mcmlxxix.