More modopts, math, comments. + Randomization
Major changes. I added into modopts.ini the ability to shut off the
binary animation (BitBomb). However, you will still get the Bitcoin
symbols. The stallwithmath function gets a FOR loop that slows it
down more effectively, and I don't see ANSI drawing errors anymore
with Syncterm or NetRunner. Better comments here and there. 
Some error handling in case the modopts.ini section is missing 
entirely or the section is there but the options are blank. Basically
bitcoin.js will default to USD and turn the BitBomb animation on. 
Also the start of trying to randomize the extra exchange rates that
are on the top line. I didn't like the fact that you had to look at the
same exact currencies all the time, when there were so many to 
look at. If you had a very large terminal screen you would see them
all, but if we randomize it, you can get different information every
time you run the door.