Finish fixing randomization. Additional Bug Fixes.
Finished fixing the randomization of the extra exchange rates at the top. 
Went through the code and caught some bug fixes that needed to be done.
Better naming of variables, because "pahkthecah" is only funny to people
in New England. Minimum terminal rows for the caller went from 11 to 12
because it looks better with the Bitcoin symbol written over the BitBomb
animation. It's honestly just aesthetics at that point. Without the animation,
you could probably make the minimum 10 or even 9 and still get something
readable where the ANSI cursor movements are not writing over each other, 
but you get a better drawn chart at 11 or higher. 12 is just going to look even
better anyway, and seriously no one should be using that few rows anyway. 
Better comments in a few places. This is fully baked. I worked on it for about 
2 months, and now I'm happy with it.