Fix file name in error msg
Fixed a holdover from when I was calling the JavaScript file
coinbase.js instead of bitcoin.js. Also, this is an opportunity
to test out GitHub webhooks integration with a new script
from echicken where this message will post to my BBS in
the General section of the local message boards.